% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{clean.lm}
\alias{clean.lm}
\title{A function to summary linear model regression set table}
\usage{
clean.lm(
  model,
  pType = "withPValue",
  estimateType = "betase",
  digits = 2,
  pDigits = 4,
  ...
)
}
\arguments{
\item{model}{linear model passed to the function}

\item{pType}{whether put P into final table, default to withPValue, another selection is withMark}

\item{estimateType}{Linear regression table type, defaults to "betase"}

\item{digits}{Digit number of the estimate}

\item{pDigits}{Digit number of the P value}

\item{...}{other parameters to pass into tidy function from broom}
}
\value{
regression table with reference level added
}
\description{
This function allows you to generate linear regression table
}
\examples{
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
clean.lm(model)
}
\keyword{Model}
